Working on Rendering
Minor changes to App.java, the parser. Added dependencies for Jzy3D, the
libraries I'll be using to render a chart based on the data the parser
processes. Added a new package with Main.java, the file to run to create
a chart. Currently, the chart is a placeholder and appears empty.