Interactivity Update
Added Jzy3D's supported interactivity through the AnalysisLauncher
class, allowing the graph to be rotated, scaled, and screenshot at will.
Added a new window to display the names of all points on the graph as
buttons with the names of all files they connect to displayed to the
side. Mousing over the button will highlight the point. Currently
working on: onClick function for the buttons to display a more advanced
view of an individual file, similar to the one described in the mockup.