Near Completion
Full functionality of the program is present with the exception of a
single serious bug. In the modeler's file detail windows, the model of
the file itself is unbounded and based on linecount. Files with a
linecount above a certain number (around 400 on a standard 1920x1080
monitor) will overflow past the screen's height, making the lower part
of the file and the information stored there totally inaccesible to the
user.