Trimming Some Fat
This update mainly removes JSON files that were used to test the program and show it in action. It also sets what version of Java should be used to run the program through Maven's pom.xml to cut down on possible errors and sets the DEBUG constants in both App.java and Main.java to false.